Refactor: useWindowSize로 리픽토링
